requirements.txt
streamlit
gTTS
playsound
speechrecognition
google-generativeai
audio-recorder-streamlit
python-dotenv
